[aim-mapping] Fix policy-classifier direction impl
When mapping the GBP policy-classifier direction
to AIM filters, the AIM filters were not being
added to the correct filter directions in the
AIM ContractSubjects. When the direction is "in"
the forward filter should be added to the "in"
bucket, the corresponding reverse filter should be
added to the "out" bucket of the ContractSubject,
and vice versa with the direction is "out". Both
filters get added to both buckets when the direction
is "bi".

There was also an issue when the direction was updated
in the policy-classifier. The updated ContractSubjects
(as a result of the policy-classifier update) were not
being written back, hence the direction update was not
taking effect. This is also being fixed.

Change-Id: I39eef396893eec98b1da4ce42f29f0b39e8ea494