[WIP]:Optimize request_endpoint_details RPC
The current request_endpoint_details needs optimization. Most
of the optimization involves reducing the number of requests
to the DB. This is done by making use of SQLAlchemy directly,
bypassing the neutron methods used to get resources.

Note that these optimizations circumvent the following code-paths:
1) RBAC and "shared" hooks for resources aren't called.

The justification for bypassing the above is that RPCs are considered
an authenticated and therefore "safe" code path. In addition, the
call to get_bound_port_context is only made if it's determined that
port-binding is neeeded. This means that case is actually made
slower due to the fact that it's not the normal path.

Change-Id: I131c423b489bfcde1185932e65529331b58db31b