[AIM] Bulk extension for ML2Plus
Improve performance of bulk read operations when using the
AIM driver by implementing a bulk extension mechanism.

Changing the extension driver framework, we can build extension
calls that have visibility over all the networks that need to be
extended, and thus can run bulk queries against the database to
reduce the number of round-trips to a constant.

Every query will still be slower as the number of objects grow,
but with this we cut down the round-trip overhead.

Change-Id: Ib884668228a5dc9be48477ca2245d762cd41bcc0