[AIM] Improve validation/repair/migration tool UTs
This change improves the validation/repair/migration tool UTs, but
does not change the system behaviour. It fixes _test_aim_resource to
verify that the tested AIM resource actually exists, and to not test
validation/repair of resource deletion or modification if the AIM
resource is monitored. It also fixes
_test_preexisting_external_network to use the proper VRF
identity. Finally, it adds code to network tests to verify that the
apic:distinguished_names attribute contains the proper DNs after
repair.

Change-Id: I1a31fde7786601c1ccfaa3df472ccdd485e8ce43