Fix Enum names for Migration scripts
Postgresql and possibly other databases require all enums
in the migration action scripts to be named appropriately.
If not errors of "unnamed Enum" occur while running migration.
gbp-db-manage gave similar errors while running with postgresql
backend. This commit is attempting to fix the errors by naming
the Enums as required.

Change-Id: Id931a9e920c4d9da9f1078a500b36dd6c1608ef4