[AIM] Repair missing extension data when validating
If extension data records for networks or subnets are missing, the
gbp-validate tool will now fail, or attempt to repair them if repair
is enabled. For external Neutron networks, the DN of the ACI
ExternalNetwork must be supplied via the migrate_ext_net_dns configuration
option in the ml2_apic_aim group, which is a dictionary keyed by the
network_id.

Change-Id: I427596b9ea01140af44e9eaf415c82935c668ebe