[AIM] Fix isomorphic address scope issues
When an isomorphic pair of address scopes reference the same
non-pre-existing VRF, set the VRF's display name
deterministically. Previously, the VRF's display name would depend on
the order in which the scopes were created or updated, which could
potentially fail validation.

If both scopes have different names that are not empty, the VRF's
display name is a combination of the v4 scope name and the v6 scope
name, seperated by a '-'. If the scopes have identical names, that
name is used as-is. If one scope's name is empty and the other's is
not, the non-empty name is used.

Another issue with validation of isomorphic address scopes is also
fixed. If the existing AddressScopeMapping record indicated that the
VRF was owned (i.e. not pre-existing), the contents of that mapping
record was ignored and was "repaired" using the normal mapping
rules. This would result in each scope in a previously isomorphic pair
having its own unique VRF rather than sharing the same VRF is
intended. Existing AddressScopeMapping records are now correctly
handled during validation.

Finally, the _check_dn() method used in many apic_aim UTs (including,
but not limited to those for address scopes) now makes sure that the
AIM resource referenced by the DN actually exists and that it has the
expected display name. This UT enhancement did not turn up any issues
unrelated to isomorphic address scopes, but should prevent such issues
from being introduced in future changes.

(cherry picked from commit 8c5830d8c96186c357ad58f1aba76a9366f8c987)

Conflicts:
	gbpservice/neutron/tests/unit/services/grouppolicy/test_aim_validation.py

Change-Id: Ia3e0cec16c0181fe100d999a3cc741d272d2008d