[AIM] Fix legacy SNAT migration
Avoid issues caused by missing NetworkMapping DB records when cleaning
up the legacy plugin's SNAT-related resources. The test_legacy_cleanup
UT is enhanced to create a router and an interfaced internal network
and to delete all mapping and extension records, so that code
manipulating legacy SNAT resources with missing mapping records is
exercised.

Missing NetworkMapping and AddressScopeMapping records are silently
ignored in delete_*_precommit methods. They are logged and ignored in
extend_*_dict methods.

When an AddressScopeMapping or a NetworkMapping record is added, the
corresponding AddressScope or Network instance's aim_mapping
relationship field is updated to reference it. This avoids aim_mapping
being None when the AddressScope or Network instance is subsequently
retrieved from the session cache within the same transaction.

Change-Id: I84f3c401df7c79deb9151b0b232262139fca9429