[AIM] Use baked queries
Use sqlalchemy's "baked query" feature to reduce the CPU time spent
building DB queries in the AIM mechanism, policy, extension, and SFC
drivers. The GBP plugin's driver-independent queries are not modified
in this patch.

To simplify initialization of the sqlalchemy bakery object, and to
share the same bakery instance as widely as possible, the various
drivers and the UT code no longer instantiate their own instances of
the DbMixin or ExtensionDbMixin classes, but instead use the mechanism
driver class instance that has these mixed into it. The GBP policy
driver class hierarchy is also adjusted slightly in order to use the
mechanism driver's bakey.

Change-Id: I5aabb1a9662ec08d6600dc5ad7e173f8ce408df3