[WIP] Use baked queries
So far, just bake address scope queries in the db and extension_db
modules.

To simplify initialization of the bakery and share the same bakery as
widely as possible, various drivers and the UT code no longer
instantiate their own instances of the DbMixin or ExtensionDbMixin
classes, but instead use the mechanism driver class that has these
mixed into it.

Change-Id: I5aabb1a9662ec08d6600dc5ad7e173f8ce408df3