[WIP] Use baked queries
So far, just bake address scope queries in db module.

To simplify initialization of the bakery and share the same bakery as
widely as possible, the UT code no longer instantiates its own
instances of the DbMixin or ExtensionDbMixin classes, but instead uses
the mechanism driver class that has these mixed into it.

Change-Id: I5aabb1a9662ec08d6600dc5ad7e173f8ce408df3