[WIP] Use baked queries
So far, bake queries in the db, extension_db, and mechanism_driver
modules, except for validation queries. Still need to bake queries in
PD modules.

To simplify initialization of the bakery and share the same bakery as
widely as possible, various drivers and the UT code no longer
instantiate their own instances of the DbMixin or ExtensionDbMixin
classes, but instead use the mechanism driver class that has these
mixed into it.

Change-Id: I5aabb1a9662ec08d6600dc5ad7e173f8ce408df3