Add support for CIDRs in allowed-adddress-pairs
The current support for allowed-address-pairs only covers
host addresses. This patch adds support for using CIDRs in
allowed address pairs. Since CIDRs are already supported in
upstream neutron, this patch only affects the backend
implementation, which is the get_gbp_details RPC and the
port notifications.

The patch also addresses cleanup of the IP => Port ID
mapping table. Entries in the table weren't affected
if an existing port's allowed-address-pairs property
was changed. This patch checks to see if the changes
address any addresses in the table, and if needed,
removes them.

Change-Id: I7bcfee46b0239577c500f2fa970e4c8400c968d0