[AIM] Move code from policy driver to mechanism driver
Move the nova_client module from the aim_mapping policy driver to the
apic_aim mechanism driver. The HAIP model class and code are not moved
in this back-port because they are shared with the legacy drivers in
stable/newton. Moving (actually copying) the HAIP code (but not model
class) to the mechanism driver may be addressed in a later patch.

(cherry picked from commit 389740089cc8010bcc82eeceb8a2e07b106a7800)
Conflicts:
	gbpservice/neutron/db/all_models.py
	gbpservice/neutron/plugins/ml2plus/drivers/apic_aim/rpc.py
	gbpservice/neutron/services/grouppolicy/drivers/cisco/apic/aim_mapping_rpc.py
	gbpservice/neutron/tests/unit/plugins/ml2plus/test_db_apic_aim.py

Change-Id: Ie1798dbe07d9da8206a202ab15c4982342855823