[AIM] Remove raw_sql RPC implementation
Remove the raw_sql RPC implementation and the unit test variations
that test it.

A planned update to this patch will remove the cache and mapping
parameters passed around to support the raw_sql implementation.

A follow-on patch will switch to the new RPC implementation as the
default, with the ablity to configure the original implemenation if
needed.

The final patch in the series will remove the original RPC
implementation and all its supporting code that is no longer needed.

Change-Id: Idc6166e19bc1f93b0bd0bbb0e5f70973cdb47a38