[AIM] Remove raw_sql RPC implementation
Remove the raw_sql RPC implementation and the unit test variations
that test it. Also, add a few REVISIT comments regarding interactions
between the mechanism and policy drivers.

A follow-on patch will switch to the new RPC implementation as the
default, with the ablity to configure the original implemenation if
needed.

The final patch in the series will remove the original RPC
implementation and all its supporting code that is no longer needed.

Change-Id: Idc6166e19bc1f93b0bd0bbb0e5f70973cdb47a38
(cherry picked from commit 90ce382e9623d01ea191d3d071edaa6430691584)