[AIM] Default to new streamlined RPC implementation
Remove the raw_sql RPC implementation and its unit tests.

Planned updates to this patch will remove the cache and mapping
parameters passed around to support the raw_sql implementation, and
switch to the new RPC implementation as the default, with the ablity
to configure the original implemenation if needed.

Change-Id: Idc6166e19bc1f93b0bd0bbb0e5f70973cdb47a38