[AIM] Use new streamlined RPC implementation be default
The default value for enable_new_rpc is now True, but it can be set to
False to use the original RPC implementation. The RPC unit tests now
dispatch directly to the mechanism driver when enable_new_rpc is True,
and directly to the policy driver when it is False.

Change-Id: Ic94da1fb569d02234fafc49b645d83f3ae5f4b49