[AIM] Improve validation output for missing external VRF
Previously, if an external network's AIM L3Outside referenced a VRF
that did not exist, validation would fail with a message that an
exception occurred, but with no indication of what caused the
problem. Now, the backtrace is logged when an exception is caught, a
specific error is reported if any of the resources needed for the
NetworkMapping DB record are missing, and a very specific error is
reported if the external VRF is missing.

Also, a requirement on acitoolkit is added to test-requirements.txt to
ensure that acitoolkit's correct branch is used, avoiding its master
branch's requirement on deepdiff, whose latest version no longer
supports python 2.

Change-Id: I7955f8e77633d9662a629c8c0628b128be3ae546