Fix I18N deprecation warnings
The gbpservice._i18n._ translation marker function is imported where
needed, and use of the _ function for log messages is removed. This
reduces the number of unique DeprecationWarning messages output during
the master branch py27 tox job from 272 to 17.

The remaining warnings are due to the use of deprecated keystone
authentication APIs, the use of the deprecated tenant_id and username
keys in API access policy files, and the use of tenant instead of
project_id within this project, and due to various issues in other
projects, including aci-integration-module, networking_sfc, and
python-opflex-agent.

Change-Id: I1cde26674ba368e239be299b2d130db98f2c0015